+++ app/components/ModalWindow.vue
@@ -7,11 +7,9 @@
             class="absolute top-0 left-0 w-full h-full bg-black/50"
             @click="open = false"
         />
-        <div
-            class="relative w-full max-h-full py-8 overflow-auto pointer-events-none"
-        >
+        <div class="relative w-full max-h-full py-8 overflow-auto">
             <div
-                class="relative w-full max-w-lg mx-auto bg-white dark:bg-slate p-8 pointer-events-auto rounded-lg text-gray text-center"
+                class="relative w-full max-w-lg mx-auto bg-white dark:bg-slate p-8 rounded-lg text-gray text-center"
             >
                 <div
                     class="absolute top-0 right-0 w-12 transform sm:translate-x-1/2 -translate-y-1/2"