+++ docs/app/app.vue
@@ -1,0 +1,8 @@
+<template>
+  <div>
+    <NuxtRouteAnnouncer />
+    <AppHeader />
+    <NuxtPage />
+    <AppFooter />
+  </div>
+</template>