+++ docs/app/components/ApiNav.vue
@@ -1,0 +1,38 @@
+<template>
+  <div
+    class="w-64 sm:w-72 px-8 py-7 sm:px-12 sm:py-10 bg-white rounded-3xl shrink-0 sticky top-6 left-0"
+  >
+    <ul>
+      <li>
+        <NuxtLink class="hover:text-secondary" to="/api/"
+          >Authentification</NuxtLink
+        >
+      </li>
+      <li>
+        <NuxtLink class="hover:text-secondary" to="/api/boards"
+          >Boards</NuxtLink
+        >
+      </li>
+      <li>
+        <NuxtLink class="hover:text-secondary" to="/api/board">Board</NuxtLink>
+      </li>
+      <li>
+        <NuxtLink class="hover:text-secondary" to="/api/areas">Areas</NuxtLink>
+      </li>
+      <li>
+        <NuxtLink class="hover:text-secondary" to="/api/area">Area</NuxtLink>
+      </li>
+      <li>
+        <NuxtLink class="hover:text-secondary" to="/api/cards">Cards</NuxtLink>
+      </li>
+      <li>
+        <NuxtLink class="hover:text-secondary" to="/api/card">Card</NuxtLink>
+      </li>
+      <li>
+        <NuxtLink class="hover:text-secondary" to="/api/comment"
+          >Comment</NuxtLink
+        >
+      </li>
+    </ul>
+  </div>
+</template>