+++ docs/app/components/ContentWrapper.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="grow shrink">
+    <slot />
+  </div>
+</template>