+++ docs/app/components/PageHero.vue
@@ -1,0 +1,28 @@
+<template>
+  <section>
+    <div class="container">
+      <div
+        class="relative bg-white rounded-3xl px-8 sm:px-12 py-10 sm:py-16 overflow-clip"
+      >
+        <div class="max-w-2xl mx-auto text-center">
+          <p class="text-secondary sm:text-lg">
+            Self-hosted Kanban boards for teams
+          </p>
+          <h1 class="text-primary text-4xl sm:text-5xl mb-4">
+            Collaborate in teams or with customers on your own server
+          </h1>
+          <p class="sm:text-lg mb-5">
+            LocalBoards is an open-source (MIT License), self-hosted Kanban
+            board system. It allows users to create boards, invite
+            collaborators, and manage Kanban cards.
+          </p>
+          <NuxtLink
+            to="/docs/"
+            class="inline-block bg-secondary text-white px-5 sm:px-6 py-2 sm:py-3 sm:text-lg rounded-full"
+            >Get started</NuxtLink
+          >
+        </div>
+      </div>
+    </div>
+  </section>
+</template>